galileo_project/requirements.txt
promptquality==0.28.0
openai==1.35.0
python-dotenv==1.0.0
pandas==2.2.2
